requirements.txt
torch
gensim
faiss-cpu
fastapi
uvicorn
pandas
numpy
jupyter
notebook
ipykernel